backend/requirements.txt
fastapi[standard]
requests
python-dotenv
supabase
mailgun
websockets